fix: address Copilot review feedback on PR #155
- Move parseSimpleYaml to src/lib/yaml.ts to decouple from WeaveService
- Add maxOutputLength (50MB) to gunzipSync to prevent OOM on crafted bundles
- Fix extractYamlBlockScalar to preserve relative indentation in block scalars
- Remove unused extractAgentMetadata import from tests
